requirements.txt
streamlit
numpy
pandas
langchain>=0.1.0
langchain-google-genai>=1.0.3
langchain-community
python-dotenv
google-generativeai>=0.5.0